Add Skia clean skip
Now we don't need to keep that line commented out! Whether this lasts for very long before we remove Skia entirely remains to be seen, but it'll be useful for a sliver of time!

Stage whisper: I used ChatGPT to write this. Bash scares me.

Diffs=
64ff27de26 Add Skia clean skip (#8650)

+++ kotlin/src/main/cpp/clean_all.sh
@@ -9,6 +9,20 @@ fi
 
 RUNTIME_ANDROID_HOME=$PWD/../../../..
 
+### Parse arguments
+
+SKIP_SKIA=false
+
+while [[ "$#" -gt 0 ]]; do
+  case $1 in
+    --skip-skia|-s) SKIP_SKIA=true ;;
+    *) echo "Warning: Unknown parameter passed: $1. Ignoring. Usage: $0 [--skip-skia|-s]" >&2 ;;
+  esac
+  shift
+done
+
+### End parse arguments
+
 if [ -z "${RIVE_RUNTIME_DIR}" ]; then
   echo "RIVE_RUNTIME_DIR is not set"
   if [ -d "${RUNTIME_ANDROID_HOME}/submodules/rive-runtime" ]; then
@@ -38,7 +52,7 @@ pushd "$RIVE_RUNTIME_DIR"/skia/renderer/
 rm -rf ./build/android
 popd
 
-# PLS
+# Rive Renderer
 pushd "$RIVE_RUNTIME_DIR"/renderer
 rm -rf ./out
 rm -rf ./android
@@ -47,6 +61,8 @@ rm -rf ./obj
 popd
 
 # Skia
-pushd "$RIVE_RUNTIME_DIR"/skia/dependencies/
-rm -rf ./skia ./skia_debug
-popd
+if [ "$SKIP_SKIA" = false ]; then
+  pushd "$RIVE_RUNTIME_DIR"/skia/dependencies/
+  rm -rf ./skia ./skia_debug
+  popd
+fi
